% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{aggregate_note}
\alias{aggregate_note}
\title{Aggregate Contents of a Note}
\usage{
aggregate_note(selected_doc_id, annotations, cui_elements)
}
\arguments{
\item{selected_doc_id}{Document ID for the note to which the sentence belongs.}

\item{annotations}{NLP annotations dataframe.}

\item{cui_elements}{Vector of UMLS concept unique identifier (CUI) elements derived from the search query.}
}
\value{
Aggregated note in one text string.
}
\description{
When using atomized notes, this function 'pastes' back the different sections together in the intended order. Preselected lemmas are marked, along with those for which thr CUI is
in the list of interest.
}
\keyword{internal}
