% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{aggregate_paragraph}
\alias{aggregate_paragraph}
\title{Aggregate Contents of a Paragraph}
\usage{
aggregate_paragraph(df_index, annotations, sentences_df)
}
\arguments{
\item{df_index}{Index of sentences dataframe; allows iteration.}

\item{annotations}{NLP annotations dataframe.}

\item{sentences_df}{Dataframe of unique sentences.}
}
\value{
Aggregated paragraph in one text string.
}
\description{
When using atomized notes, this function 'pastes' back the different sections together in the intended order for ONE paragraph.
There is NO marking of lemmas.
}
\keyword{internal}
