% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{colorize}
\alias{colorize}
\title{Highlight Keywords and Concepts}
\usage{
colorize(get_output)
}
\arguments{
\item{get_output}{Dataframe with field 'selected', the latter containing text string corresponding to selected sentences.}
}
\value{
Text string with HTML markup.
}
\description{
Adds color and changes font to bold for selected strings of sentences and notes. The current method uses HTML.
}
