% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{complete_case}
\alias{complete_case}
\title{Mark a Case as Completed}
\usage{
complete_case(uri_fun, user, password, host, port, database, selected_patient)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{selected_patient}{Selected patient.}
}
\description{
Once all sentences before a recorded event have been annotated by the end user, this functions marks patient record as 'reviewed'.
}
\keyword{internal}
