% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_project}
\alias{create_project}
\title{Create a New CEDARS Project}
\usage{
create_project(
  uri_fun,
  user,
  password,
  host,
  port,
  database,
  project_name,
  investigator_name
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{project_name}{Research or QA project name.}

\item{investigator_name}{Investigator name.}
}
\description{
Creates a new MongoDB database and collections needed for a CEDARS annotation project. The MongoDB account used must have sufficient privileges.
}
\examples{
\dontrun{
create_project(uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234',
host = 'server1234', port = NA, database = 'TEST_PROJECT', project_name = 'Test project',
investigator_name = 'Dr Smith')
}
}
