% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_download}
\alias{db_download}
\title{Obtain Patient Notes}
\usage{
db_download(uri_fun, user, password, host, port, database, patient_id)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{patient_id}{Patient ID for which notes are being requested.}
}
\value{
Dataframe of full notes and/or note parts with associated metadata.
}
\description{
Downloads the notes for one patient. If there are missing fields, those are created with NA values inserted.
}
