% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Load Options on Startup}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name.}

\item{pkgname}{Package name.}
}
\description{
Disables scientific notation which can be a problem for large ID's.
}
\keyword{internal}
