% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{download_filtered_tags}
\alias{download_filtered_tags}
\title{Download Notes Metadata Summary}
\usage{
download_filtered_tags(
  uri_fun,
  user,
  password,
  host,
  replica_set,
  port,
  database,
  eval_query = TRUE
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB server host.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{eval_query}{Should tags be evaluated against stored CEDARS query?}
}
\value{
{
List containing one dataframe per text_tag_x of interest, indicating selected status.
Can take time to download for large datasets.
}
}
\description{
Downloads text_tag_x info for the metadata tags, along with "include" and/or
"exclude" criteria in the core CEDARS search query.
}
\examples{
\dontrun{
download_filtered_tags(uri_fun = mongo_uri_standard, user = 'John',
password = 'db_password_1234', host = 'server1234', replica_set = NA, port = NA,
database = 'TEST_PROJECT')
}
}
