% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{download_outcomes}
\alias{download_outcomes}
\title{Download Event Data}
\usage{
download_outcomes(uri_fun, user, password, host, database)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{database}{MongoDB database name.}
}
\description{
Downloads patient event data. Typically done after all records have been annotated and the project is complete.
}
\examples{
\dontrun{
download_outcomes(uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234',
host = 'server1234')
}
}
