% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMLS_processor.R
\name{fast_merge_umls}
\alias{fast_merge_umls}
\title{Merge UMLS Tags}
\usage{
fast_merge_umls(query_x, umls_selected, by_x, n_grams)
}
\arguments{
\item{query_x}{Vector of tokens to match.}

\item{umls_selected}{UMLS dictionary.}

\item{by_x}{Field name for tokens to match.}

\item{n_grams}{Length of tokens to match.}
}
\value{
Tokens matched with a CUI.
}
\description{
Fast implementation of matching process to find UMLS concept unique identifiers (CUI's) corresponding to tokens from EHR documents.
}
