% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{find_model}
\alias{find_model}
\title{Find NLP Model}
\usage{
find_model(selected_model_path)
}
\arguments{
\item{selected_model_path}{Initial selected model path, if NA we find one.}
}
\value{
Model path.
}
\description{
Finds NLP model to use, if not specified. If no model present, we download the default.
If several present, by default we use the first one by alphabetical order.
}
\examples{
\dontrun{
find_model(NA)
}
}
\keyword{internal}
