% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{format_keywords}
\alias{format_keywords}
\title{Format Keyword Elements}
\usage{
format_keywords(keyword_elements)
}
\arguments{
\item{keyword_elements}{Keywords.}
}
\value{
Formatted keywords.
}
\description{
Prepares keyword element vector, converting wildcards to Regex syntax. Wildcards include '?' for any single character and '*' for any number of characters including zero.
}
