% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{get_model}
\alias{get_model}
\title{Get a NLP Model}
\usage{
get_model(model_name = "english-ewt", platform = "udpipe")
}
\arguments{
\item{model_name}{Name of models to download.}

\item{platform}{Name of NLP platform, currently only UDPipe is supported.}
}
\value{
Saves model in inst/models.
}
\description{
Downloads a NLP model, presently only UDPipe models supported.
}
