% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negex.R
\name{id_expander}
\alias{id_expander}
\title{Expand ID's}
\usage{
id_expander(index, work_df, field)
}
\arguments{
\item{index}{Row position within annotations dataframe.}

\item{work_df}{Working annotations dataframe.}

\item{field}{Field to duplicate.}
}
\value{
Series of duplicated field values.
}
\description{
Duplicates paragraph or sentence ID's to help process negation based on negated token series.
}
