% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{initialize_notes}
\alias{initialize_notes}
\title{Initialize EHR Notes}
\usage{
initialize_notes(uri_fun, user, password, host, database)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{database}{MongoDB database name.}
}
\description{
Deletes all patient notes from the database.
}
\examples{
\dontrun{
initialize_notes(uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234',
host = 'server1234', database = 'TEST_PROJECT')
}
}
