% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{initialize_users}
\alias{initialize_users}
\title{Initialize End User List}
\usage{
initialize_users(uri_fun, user, password, host, port, database)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}
}
\description{
Deletes all CEDARS end user credentials information.
}
\examples{
\dontrun{
initialize_users(uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234',
host = 'server1234', port = NA, database = 'TEST_PROJECT')
}
}
