% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{lemma_match}
\alias{lemma_match}
\title{Match Lemmas to Regex Keywords}
\usage{
lemma_match(annotations, keyword_elements)
}
\arguments{
\item{annotations}{Patient-specific NLP annotations.}

\item{keyword_elements}{Keywords with Regex syntax.}
}
\value{
Annotations dataframe with one extra column.
}
\description{
Matches lemmas in NLP annotations to keywords, considering Regex syntax.
}
