% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{lock_records}
\alias{lock_records}
\title{Lock Record for one Patient}
\usage{
lock_records(
  uri_fun,
  user,
  password,
  host,
  database,
  end_user,
  patient_id = NA
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{database}{MongoDB database name.}

\item{end_user}{CEDARS end user name.}

\item{patient_id}{Used if a specific patient record is requested, instead of a search for next record to annotate.}
}
\description{
Locks record for selected patient. The first available unreviewed, unlocked record is locked for the CEDARS end user.
}
