% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{lock_records_admin}
\alias{lock_records_admin}
\title{Lock Record for Admin Work
When tasks need to be performed by the CEDARS server, this functions locks one patient record at a time. The record is locked only if not already locked by a CEDARS end user. Admin lock and end user lock do not coexist.}
\usage{
lock_records_admin(
  uri_fun,
  user,
  password,
  host,
  replica_set,
  port,
  database,
  patient_id
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{patient_id}{Patient ID being locked.}
}
\description{
Lock Record for Admin Work
When tasks need to be performed by the CEDARS server, this functions locks one patient record at a time. The record is locked only if not already locked by a CEDARS end user. Admin lock and end user lock do not coexist.
}
\keyword{internal}
