% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{mark}
\alias{mark}
\title{Mark Tokens}
\usage{
mark(annotations, cui_elements)
}
\arguments{
\item{annotations}{NLP annotation dataframe.}

\item{cui_elements}{Vector of search query CUI's.}
}
\value{
Full sentence with marked tokens.
}
\description{
Marks tokens corresponding to keywords/CUI's so that they can be highlighted later.
}
