% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{mongo_connect}
\alias{mongo_connect}
\title{Connect to MongoDB}
\usage{
mongo_connect(uri_fun, user, password, host, port, database, mongo_collection)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{mongo_collection}{MongoDB collection; if NA, will connect to DB itself.}
}
\description{
Basic function used by CEDARS to connect to MongoDB instance. Wraps mongo() from mongolite.
}
