% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negex_data.R
\docType{data}
\name{negex}
\alias{negex}
\title{Negex data
Apache License 2.0}
\format{
An object of class \code{'data.frame'}.
}
\source{
\href{https://code.google.com/archive/p/negex/downloads}{Google Code Archive}
}
\usage{
data(negex)
}
\description{
Negex data
Apache License 2.0
}
\examples{
data(negex)

}
\references{
Chapman et al. (2013) Stud Health Technol Inform 192:677-681
(\href{https://pubmed.ncbi.nlm.nih.gov/23920642/}{PubMed})
}
\keyword{datasets}
