% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negex.R
\name{negex_processor}
\alias{negex_processor}
\title{Annotate for Negation}
\usage{
negex_processor(annotated_text, negex_simp, negex_depth = 6)
}
\arguments{
\item{annotated_text}{Dataframe of NLP annotations.}

\item{negex_simp}{Dataframe of simplified negex.}

\item{negex_depth}{Maximum distance between word to label and negation term, before or after. Default is 6, as per original paper by Chapman et al.}
}
\value{
Dataframe with added negation information.
}
\description{
Annotates EHR text for negation, based on Negex, as simple medical negation lexicon. Tokens are labelled by proximity. The original paper went up to 6 tokens before or after negation.
}
\keyword{internal}
