% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negex.R
\name{negex_token_tagger}
\alias{negex_token_tagger}
\title{Compute Token Series}
\usage{
negex_token_tagger(index, work_df, negex_depth)
}
\arguments{
\item{index}{Position of negation item.}

\item{work_df}{Working dataframe of NLP annotations.}

\item{negex_depth}{Maximum distance from index to which negation will spread.}
}
\value{
Series of token positions within the working dataframe.
}
\description{
Computes negated token series as a function of index, i.e. position of negation item.
}
