% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{negex_upload}
\alias{negex_upload}
\title{Upload NegEx}
\usage{
negex_upload(udmodel_path, uri_fun, user, password, host, database)
}
\arguments{
\item{udmodel_path}{Path to NLP model file.}

\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{database}{MongoDB database name.}
}
\description{
Prepares and uploads NegEx negation lexicon. It is not absolutely required for CEDARS to function but in practice will improve search accuracy for most applications.
}
\examples{
\dontrun{
negex_upload(udmodel_path = 'models/english-ewt-ud-2.4-190531.udpipe',
uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234', host = 'server1234',
database = 'TEST_PROJECT')
}
}
