% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse_query}
\alias{parse_query}
\title{Parse a Search Query}
\usage{
parse_query(search_query)
}
\arguments{
\item{search_query}{Medical corpus query containg keywords/CUI's, boolean elements and other operators ('AND', 'OR', '!', '(', or ')').}
}
\value{
List containing the query vector, operator mask, CUI mask, keyword mask, keyword elements and CUI elements.
}
\description{
Extracts pertinent elements from search query.
}
\keyword{internal}
