% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{password_verification}
\alias{password_verification}
\title{Verify CEDARS End User Credentials}
\usage{
password_verification(
  uri_fun,
  user,
  password,
  host,
  port,
  database,
  end_user,
  end_user_password
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{end_user}{CEDARS end user name.}

\item{end_user_password}{CEDARS end user password.}
}
\value{
TRUE for correct credentials, FALSE for incorrect.
}
\description{
Verifies CEDARS end user name and password against information stored in MongoDB. Each CEDARS project is independent and occupies a different DB with its own user credentials collection.
}
\keyword{internal}
