% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{paste_sections}
\alias{paste_sections}
\title{Paste Individual Note Sections}
\usage{
paste_sections(section_index, note_list)
}
\arguments{
\item{section_index}{Index number for the section within the note.}

\item{note_list}{List of dataframes, each one containing NLP annotations for a note section.}
}
\value{
Vector of pasted sections.
}
\description{
Pastes the notes sections within a list.
}
\keyword{internal}
