% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{patient_roster_update}
\alias{patient_roster_update}
\title{Update Patient List}
\usage{
patient_roster_update(
  uri_fun,
  user,
  password,
  host,
  replica_set,
  port,
  database,
  patient_vect = NA
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{patient_vect}{Vector of patient ID's. Default is NA, in which case all unlocked records will be listed.}
}
\description{
Updates the 'PATIENTS' collection with new entries. Lists all patients who have at
least one note.
}
\keyword{internal}
