% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{populate_annotations}
\alias{populate_annotations}
\title{Prepare Database for NLP Annotation}
\usage{
populate_annotations(
  uri_fun,
  user,
  password,
  host,
  replica_set,
  port,
  database
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}
}
\description{
Creates 'ANNOTATIONS' and 'PATIENTS' collections, assuming main DB collection and notes were already set up.This is necessary before launching the NLP annotation process.
}
\keyword{internal}
