% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{populate_notes}
\alias{populate_notes}
\title{Prepare Database for Patient Notes}
\usage{
populate_notes(uri_fun, user, password, host, database)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{database}{MongoDB database name.}
}
\description{
Sets up the 'NOTES' collection which will hold original EHR documents. Those notes can exist as complete documents (one doc per note) or in atomized form (several docs, each a subsection of one note).
}
