% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{populate_query}
\alias{populate_query}
\title{Prepare Database to Receive the Search Query}
\usage{
populate_query(uri_fun, user, password, host, replica_set, port, database)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}
}
\description{
Sets up the 'QUERY' collection to store the search query and associated information.
}
\keyword{internal}
