% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{populate_users}
\alias{populate_users}
\title{Prepare Database for User Credentials}
\usage{
populate_users(uri_fun, user, password, host, port, database)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}
}
\description{
Sets up 'USERS' collection where CEDARS end user names and passwords will be retained.
}
