% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{sanitize}
\alias{sanitize}
\title{Sanitize Text}
\usage{
sanitize(text)
}
\arguments{
\item{text}{Raw text from the EHR.}
}
\value{
Cleaned text in ASCII format stripped of '*' characters.
}
\description{
Removes non-ASCII characters and '*' from clinical corpora or user input.
}
