% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{sanitize_query}
\alias{sanitize_query}
\title{Sanitize Query}
\usage{
sanitize_query(query)
}
\arguments{
\item{query}{Raw text from the query.}
}
\value{
Cleaned query.
}
\description{
Removes non-ASCII characters from search query and attempts to correct formulation errors.
}
