% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{save_credentials}
\alias{save_credentials}
\title{Save MongoDB Credentials}
\usage{
save_credentials(user, password, host, database, destination_path = getwd())
}
\arguments{
\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB server host.}

\item{database}{MongoDB database name.}

\item{destination_path}{Folder where the files should be saved. Default is working directory.}
}
\description{
Saves MongoDB credentials as 'db_credentials.Rdata' and Shiny app file as 'app.R'. Those two files should be copied to the Shiny Server app directory. Needed only if using Shiny Server; credentials are entered in the command line for local app use.
}
\examples{
\dontrun{
save_credentials(user = 'John', password = 'db_password_1234', host = 'server1234',
destination_path = getwd())
}
}
