% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{sentence_search}
\alias{sentence_search}
\title{Functions which mediate keyword/CUI search on NLP-annotated corpora}
\usage{
sentence_search(parse_result, annotations, use_negation, hide_duplicates)
}
\arguments{
\item{parse_result}{Results of search query parsing.}

\item{annotations}{Patient-specific NLP annotations.}

\item{use_negation}{Should negated items be ignored in the keyword/concept search?}

\item{hide_duplicates}{Should duplicated sentences be removed for search results?}
}
\value{
All sentences matching the query.
}
\description{
Execute Search on EHR Notes for One Patient
}
\details{
Performs search for queries based on lemmas (i.e. keywords) and UMLS CUI's
}
