% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_patients_data.R
\docType{data}
\name{simulated_patients}
\alias{simulated_patients}
\title{Simulated patient data
GPL-3 license}
\format{
An object of class \code{'data.frame'}.
}
\usage{
data(simulated_patients)
}
\description{
Simulated patient data
GPL-3 license
}
\examples{
data(simulated_patients)

}
\keyword{datasets}
