% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{tag_filter}
\alias{tag_filter}
\title{Filter Dataframe Based on Metadata Tags}
\usage{
tag_filter(tagged_df, tag_query)
}
\arguments{
\item{tagged_df}{Input dataframe.}

\item{tag_query}{List of "included" and "excluded" tags.}
}
\value{
Subset of input dataframe fitting metadata inclusion/exclusion criteria.
}
\description{
Parses a dataframe, looks for text tags and applies filter based on text_tag_x fields.
If no "include" criterion specified, all rows kept unless tag listed in "exclude".
}
\keyword{internal}
