% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{terminate_project}
\alias{terminate_project}
\title{Terminate CEDARS Project}
\usage{
terminate_project(uri_fun, user, password, host, port, database, fast = FALSE)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{fast}{If TRUE, delete everything without asking security questions.}
}
\value{
{
Confirmation that requested operation was completed, or error message if attempt failed.
}
}
\description{
Everything is removed, including dictionaries. MongoDB account used must have sufficient privileges.
}
\examples{
\dontrun{
terminate_project(uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234',
host = 'server1234', port = NA, database = 'TEST_PROJECT')
}
}
