% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{terminate_project_new}
\alias{terminate_project_new}
\title{Terminate CEDARS Project - short version}
\usage{
terminate_project_new(
  uri_fun,
  user,
  password,
  host,
  replica_set,
  port,
  database
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}
}
\description{
Everything is removed, including dictionaries. MongoDB account used must have sufficient privileges.
}
\examples{
\dontrun{
terminate_project(uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234',
host = 'server1234', port = NA, database = 'TEST_PROJECT')
}
}
\keyword{internal}
