% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{token_splitter}
\alias{token_splitter}
\title{Split a Token}
\usage{
token_splitter(i, df, max_grams)
}
\arguments{
\item{i}{Iteration number, i.e. row number withint MRCONSO.}

\item{df}{MRCONSO dataframe.}

\item{max_grams}{Maximum length of token in grams.}
}
\value{
A list which includes the original data for one UMLS concept unique identifier (CUI).
}
\description{
Splits a token in its different grams and outputs. Processes one row from the UMLS MRCONSO file at a time.
}
\keyword{internal}
