% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMLS_processor.R
\name{umls_processor}
\alias{umls_processor}
\title{Tag Tokens With UMLS Concepts}
\usage{
umls_processor(annotated_text, umls_selected, max_n_grams_length)
}
\arguments{
\item{annotated_text}{Dataframe of NLP annotations.}

\item{umls_selected}{UMLS dictionary.}

\item{max_n_grams_length}{Maximum length of tokens for matching with UMLS concept unique identifiers (CUI's). Shorter values will result in faster processing. If ) is chosen, UMLS CUI tags will not be provided.}
}
\value{
Dataframe of NLP annotations with added UMLS tags.
}
\description{
Main processor for matching to UMLS concepts.
}
