% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_io.R
\name{unlock_records_admin}
\alias{unlock_records_admin}
\title{Remove Admin Lock}
\usage{
unlock_records_admin(uri_fun, user, password, host, port, database, patient_id)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{patient_id}{ID of patient record being unlocked.}
}
\description{
Once the CEDARS server has completed work on a given patient, admin lock is removed so end users can lock and access the record.
}
