% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_upload}
\alias{db_upload}
\title{Upload Annotations}
\usage{
db_upload(
  uri_fun,
  user,
  password,
  host,
  replica_set,
  port,
  database,
  patient_id,
  annotations
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{patient_id}{Patient ID.}

\item{annotations}{NLP annotations.}
}
\description{
Uploads NLP annotations for one patient.
}
\keyword{internal}
