% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{initialize_annotations}
\alias{initialize_annotations}
\title{Initialize Annotations
Deletes all NLP annotations and patient-specific information, including clinical event dates.New, empty 'ANNOTATIONS' and 'PATIENTS' collections are created. Dictionaries and original patient notes are preserved.}
\usage{
initialize_annotations(
  uri_fun,
  user,
  password,
  host,
  replica_set,
  port,
  database,
  override = FALSE
)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}

\item{database}{MongoDB database name.}

\item{override}{Override safety checks, move directly to initialization!}
}
\value{
{
Confirmation that requested operation was completed, or error message if attempt failed.
}
}
\description{
Initialize Annotations
Deletes all NLP annotations and patient-specific information, including clinical event dates.New, empty 'ANNOTATIONS' and 'PATIENTS' collections are created. Dictionaries and original patient notes are preserved.
}
\examples{
\dontrun{
initialize_annotations(uri_fun = mongo_uri_standard, user = 'John', password = 'db_password_1234',
host = 'server1234', port = NA, database = 'TEST_PROJECT')
}
}
