% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{mongo_uri_standard}
\alias{mongo_uri_standard}
\title{Prepare MongoDB URI string, most commonly used format}
\usage{
mongo_uri_standard(user, password, host, replica_set = NA, port = NA)
}
\arguments{
\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{replica_set}{MongoDB replica set, if indicated.}

\item{port}{MongoDB port.}
}
\value{
URI string.
}
\description{
Formats the MongoDB URI string for use by package mongolite. In this case the 'standard' URI format is used.
}
\examples{
\dontrun{
mongo_uri_standard(user = 'John', password = 'db_password_1234', host = 'server1234', port = NA)
}
}
