% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{patient_roster_update}
\alias{patient_roster_update}
\title{Update Patient List}
\usage{
patient_roster_update(uri_fun, user, password, host, database)
}
\arguments{
\item{uri_fun}{Uniform resource identifier (URI) string generating function for MongoDB credentials.}

\item{user}{MongoDB user name.}

\item{password}{MongoDB user password.}

\item{host}{MongoDB host server.}

\item{database}{MongoDB database name.}
}
\description{
Updates the 'PATIENTS' collection with new entries. Used after NLP annotations have been uploaded.
}
